# openapi/wallet-api.yaml
openapi: 3.1.0
info:
  title: Universal AI Wallet — Wallet API
  version: 1.0.0
  description: Account, balance, keys, and top-ups.
servers:
  - url: https://api.example.com/wallet/v1
paths:
...
  /keys:
    get:
      operationId: listKeys
      summary: List virtual keys
      tags: [Keys]
      security:
        - JwtAuth: []
      responses:
        '200':
          description: Keys
          content:
            application/json:
              schema:
                type: object
                properties:
                  data:
                    type: array
                    items:
                      $ref: '#/components/schemas/VirtualKey'

    post:
      operationId: createKey
      summary: Create virtual key
      tags: [Keys]
      security:
        - JwtAuth: []
      parameters:
        - name: Idempotency-Key
          in: header
          schema:
            type: string
            format: uuid
      requestBody:
        content:
          application/json:
            schema:
              type: object
              properties:
                name:
                  type: string
                rpm_limit:
                  type: integer
                  default: 60
                tpm_limit:
                  type: integer
                  default: 100000
      responses:
        '201':
          description: Key created — plaintext shown once
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/VirtualKeyCreated'
  /keys/{keyId}:
    delete:
      operationId: revokeKey
      summary: Revoke key
      tags: [Keys]
      security:
        - JwtAuth: []
      parameters:
        - name: keyId
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '204':
          description: Revoked
...
  /oauth/authorize:
    get:
      operationId: authorize
      summary: Authorization endpoint (consent descriptor; requires user session)
      tags: [OAuth]
      servers:
        - url: https://api.example.com
      security:
        - JwtAuth: []
      parameters:
        - name: client_id
          in: query
          required: true
          schema:
            type: string
        - name: redirect_uri
          in: query
          required: true
          schema:
            type: string
        - name: response_type
          in: query
          schema:
            type: string
            default: code
        - name: state
          in: query
          schema:
            type: string
        - name: scope
          in: query
          schema:
            type: string
        - name: code_challenge
          in: query
          schema:
            type: string
        - name: code_challenge_method
          in: query
          schema:
            type: string
            default: S256
      responses:
        '200':
          description: Consent descriptor
        '400':
          description: Invalid client or redirect_uri
        '401':
          description: User session required
  /oauth/authorize/consent:
    post:
      operationId: consent
      summary: Approve spend cap and issue an authorization code
      tags: [OAuth]
      servers:
        - url: https://api.example.com
      security:
        - JwtAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [client_id, redirect_uri]
              properties:
                client_id:
                  type: string
                redirect_uri:
                  type: string
                state:
                  type: string
                scope:
                  type: string
                code_challenge:
                  type: string
                code_challenge_method:
                  type: string
                  default: S256
                spend_limit_microdollars:
                  type: integer
                  minimum: 0
                  nullable: true
                reset_period:
                  type: string
                  enum: [monthly, lifetime]
                  default: monthly
      responses:
        '200':
          description: Redirect URI with code + state
          content:
            application/json:
              schema:
                type: object
                properties:
                  redirect_uri:
                    type: string
                  code:
                    type: string
                  state:
                    type: string
  /oauth/token:
    post:
      operationId: token
      summary: Token endpoint (authorization_code + PKCE or refresh_token)
      tags: [OAuth]
      servers:
        - url: https://api.example.com
      security: []
      requestBody:
        required: true
        content:
          application/x-www-form-urlencoded:
            schema:
              type: object
              required: [grant_type, client_id]
              properties:
                grant_type:
                  type: string
                  enum: [authorization_code, refresh_token]
                code:
                  type: string
                code_verifier:
                  type: string
                refresh_token:
                  type: string
                redirect_uri:
                  type: string
                client_id:
                  type: string
                client_secret:
                  type: string
                scope:
                  type: string
      responses:
        '200':
          description: Token set
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TokenSet'
        '400':
          description: invalid_grant / invalid_client / unsupported_grant_type
  /oauth/userinfo:
    get:
      operationId: userinfo
      summary: OIDC UserInfo
      tags: [OAuth]
      servers:
        - url: https://api.example.com
      security:
        - OAuthBearer: []
      responses:
        '200':
          description: User profile + app install context
        '401':
          description: invalid_token
...
components:
  securitySchemes:
    JwtAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
    OAuthBearer:
      type: http
      scheme: bearer
      description: OAuth2 app access token (JWT with app_install_id claim)
    PartnerAdmin:
      type: apiKey
      in: header
      name: X-Partner-Admin-Token
  schemas:
...
    VirtualKey:
      type: object
      properties:
        id:
          type: string
          format: uuid
        name:
          type: string
        key_prefix:
          type: string
        rpm_limit:
          type: integer
        tpm_limit:
          type: integer
        created_at:
          type: string
          format: date-time
        revoked_at:
          type: string
          format: date-time
    VirtualKeyCreated:
      allOf:
        - $ref: '#/components/schemas/VirtualKey'
        - type: object
          required: [key]
          properties:
            key:
              type: string
              description: Full key — only returned at creation
...
    TokenSet:
      type: object
      properties:
        access_token:
          type: string
        id_token:
          type: string
        refresh_token:
          type: string
        token_type:
          type: string
          default: Bearer
        expires_in:
          type: integer
        scope:
          type: string